REF: MultiIndex Indexing tests (#33053)
