+++ pandas/tests/indexes/multi/test_get_set.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 import pandas as pd
-from pandas import CategoricalIndex, Index, MultiIndex
+from pandas import CategoricalIndex, MultiIndex
 import pandas._testing as tm
 
 
@@ -25,90 +25,6 @@ def test_get_level_number_integer(idx):
         idx._get_level_number(2)
     with pytest.raises(KeyError, match="Level fourth not found"):
         idx._get_level_number("fourth")
-
-
-def test_get_level_values(idx):
-    result = idx.get_level_values(0)
-    expected = Index(["foo", "foo", "bar", "baz", "qux", "qux"], name="first")
-    tm.assert_index_equal(result, expected)
-    assert result.name == "first"
-
-    result = idx.get_level_values("first")
-    expected = idx.get_level_values(0)
-    tm.assert_index_equal(result, expected)
-
-    # GH 10460
-    index = MultiIndex(
-        levels=[CategoricalIndex(["A", "B"]), CategoricalIndex([1, 2, 3])],
-        codes=[np.array([0, 0, 0, 1, 1, 1]), np.array([0, 1, 2, 0, 1, 2])],
-    )
-
-    exp = CategoricalIndex(["A", "A", "A", "B", "B", "B"])
-    tm.assert_index_equal(index.get_level_values(0), exp)
-    exp = CategoricalIndex([1, 2, 3, 1, 2, 3])
-    tm.assert_index_equal(index.get_level_values(1), exp)
-
-
-def test_get_value_duplicates():
-    index = MultiIndex(
-        levels=[["D", "B", "C"], [0, 26, 27, 37, 57, 67, 75, 82]],
-        codes=[[0, 0, 0, 1, 2, 2, 2, 2, 2, 2], [1, 3, 4, 6, 0, 2, 2, 3, 5, 7]],
-        names=["tag", "day"],
-    )
-
-    assert index.get_loc("D") == slice(0, 3)
-
-
-def test_get_level_values_all_na():
-    # GH 17924 when level entirely consists of nan
-    arrays = [[np.nan, np.nan, np.nan], ["a", np.nan, 1]]
-    index = pd.MultiIndex.from_arrays(arrays)
-    result = index.get_level_values(0)
-    expected = pd.Index([np.nan, np.nan, np.nan], dtype=np.float64)
-    tm.assert_index_equal(result, expected)
-
-    result = index.get_level_values(1)
-    expected = pd.Index(["a", np.nan, 1], dtype=object)
-    tm.assert_index_equal(result, expected)
-
-
-def test_get_level_values_int_with_na():
-    # GH 17924
-    arrays = [["a", "b", "b"], [1, np.nan, 2]]
-    index = pd.MultiIndex.from_arrays(arrays)
-    result = index.get_level_values(1)
-    expected = Index([1, np.nan, 2])
-    tm.assert_index_equal(result, expected)
-
-    arrays = [["a", "b", "b"], [np.nan, np.nan, 2]]
-    index = pd.MultiIndex.from_arrays(arrays)
-    result = index.get_level_values(1)
-    expected = Index([np.nan, np.nan, 2])
-    tm.assert_index_equal(result, expected)
-
-
-def test_get_level_values_na():
-    arrays = [[np.nan, np.nan, np.nan], ["a", np.nan, 1]]
-    index = pd.MultiIndex.from_arrays(arrays)
-    result = index.get_level_values(0)
-    expected = pd.Index([np.nan, np.nan, np.nan])
-    tm.assert_index_equal(result, expected)
-
-    result = index.get_level_values(1)
-    expected = pd.Index(["a", np.nan, 1])
-    tm.assert_index_equal(result, expected)
-
-    arrays = [["a", "b", "b"], pd.DatetimeIndex([0, 1, pd.NaT])]
-    index = pd.MultiIndex.from_arrays(arrays)
-    result = index.get_level_values(1)
-    expected = pd.DatetimeIndex([0, 1, pd.NaT])
-    tm.assert_index_equal(result, expected)
-
-    arrays = [[], []]
-    index = pd.MultiIndex.from_arrays(arrays)
-    result = index.get_level_values(0)
-    expected = pd.Index([], dtype=object)
-    tm.assert_index_equal(result, expected)
 
 
 def test_set_name_methods(idx, index_names):
